\name{DIList-class}
\docType{class}

\alias{DIList}
\alias{DI2IS}
\alias{getDistance}
\alias{InteractionSet-defunct}

\title{DIList class and methods}

\description{
This class is now defunct.
}

\details{
Users should use InteractionSet objects and methods for future work.
Existing DIList objects can be converted into InteractionSet objects using the DI2IS function.
} 

\author{Aaron Lun}

\keyword{counting}
