\name{boxPairs}
\alias{boxPairs}

\title{Put bin pairs into boxes}
\description{Match smaller bin pairs to the larger bin pairs in which they are nested.}

\usage{
boxPairs(reference, ..., fragments)
}

\arguments{
\item{reference}{an integer scalar specifying the reference bin size}
\item{...}{one or more named list objects of bin pair data generated by \code{\link{squareCounts}}, with smaller bin sizes than \code{reference}}
\item{fragments}{a \code{GRanges} object containing the boundaries of the restriction fragments in the genome}
}

\value{
A list object with several components:
\item{indices}{a named list of integer vectors for every list in the ellipsis, see Details} 
\item{pairs}{a dataframe of integer \code{anchor.id} and \code{target.id} fields, describing the parent bin pair}
\item{region}{a \code{GRanges} object containing the coordinates for each parent bin}
}

\details{
Consider the bin size specified in \code{reference}. 
Pairs of these bins are referred to here as the parent bin pairs, and are described in the output \code{pairs} and \code{region}.
The function accepts a number of lists of bin pairs in the ellipsis, referred to here as input bin pairs.
The aim is to identify the parent bin pair in which each input bin pair is nested.

One vector will be present in \code{indices} for each input list in the ellipsis.
Each index in a given vector matches up to a single input bin pair in the corresponding list.
The index points to a row of \code{pairs} and indicates the parent bin pair in which that input bin pair is nested.

All input bin pair lists in the ellipsis must be constructed carefully.
In particular, the value of \code{width} in \code{\link{squareCounts}} must be such that \code{reference} is an exact multiple of each \code{width}.
This is necessary to ensure complete nesting.
Otherwise, the behaviour of the function will not be clearly defined.

Some users may wish to identify bin pairs in one list that are nested within bin pairs in another list.
This can be done by setting the larger bin size as \code{reference}, and supplying both bin pair lists to the ellipsis.
It is then a trivial matter to \code{\link{match}} the output indices for the smaller bin pairs to the output indices of the larger bin pairs.
}

\seealso{
\code{\link{squareCounts}}
}

\author{
Aaron Lun
}

\examples{
hic.file <- system.file("exdata", "hic_sort.bam", package="diffHic")
cuts <- readRDS(system.file("exdata", "cuts.rds", package="diffHic"))

# Setting up the parameters (low filter is just to get more bin pairs)
fout <- "output.h5"
preparePairs(hic.file, cuts, fout)
countPairs(fout)
y <- squareCounts(fout, cuts, width=50, filter=1)
y2 <- squareCounts(fout, cuts, width=25, filter=1)

# Clustering.
boxPairs(reference=100, larger=y, smaller=y2, fragments=cuts)
boxPairs(reference=50, larger=y, smaller=y2, fragments=cuts)

\dontshow{
unlink(fout, recursive=TRUE)
}
}
