\name{countNeighbors}
\alias{countNeighbors}

\title{Get counts for neighbouring interactions}
\description{Counts the number of read pairs for adjacent bin pairs in the interaction space.}

\usage{
countNeighbors(data, flank)
}

\arguments{
    \item{data}{a \code{DIList} object produced by \code{\link{squareCounts}}}
	\item{flank}{an integer scalar, specifying the maximum distance at which bin pairs are considered neighbors}
}

\value{
A list containing:
\item{y}{a \code{DGEList} object with one row for each bin pair in \code{data}, containing the sum of counts for the neighboring bin pairs}
\item{n}{a numeric vector for each bin pair, specifying the number of bin pairs in each neighbourhood}
}

\details{
For intra-chromosomal bin pairs, the neighbouring bin pairs are defined as those on the same diagonal in the interacion space.
For a bin pair with anchor-target indices \code{(x, y)}, neighbours are defined as all \code{(x+i, y+i)} for \code{i} in \code{[-flank, flank]}.
Some finesse is involved at the edges of the interaction space, whereby this definition is adjusted slightly to maintain the size of the neighbourhood.
In all cases, the neighbourhood is comprised of bin pairs from the same diagonal.
This avoids any differences due to the dependency of the abundance on the interaction distance.

For inter-chromosomal bin pairs, the neighbouhood is defined using a cross-like shape in the interaction space.
A bin pair is considered to be a neighbor if it has the same \code{x} and a target index in \code{[y-flank, y+flank]}, or the same \code{y} and an anchor index in \code{[x-flank, x+flank]}.
Preservation of interaction distance is not performed as distance-dependent biases are generally absent in the inter-chromosomal interaction space.
Again, some finesse is involved at the edge of the interaction space to maintain the neighbourhood size.

For each central bin pair, read pair counts are summed over all the neighboring bin pairs (with the exception of the central bin pair itself, of course). 
This can be used as a local estimate of background ligation, e.g., to call peaks in the interaction space. 
Of course, the area of the neighbourhood will not be the same as that of the central bin pair.
Abundances should be normalized to equalize the areas prior to performing comparisons.

The \code{data} object should be generated by taking the output of \code{\link{squareCounts}} with \code{filter=1}. 
Prior filtering should not be performed as low-abundance bin pairs are still informative neighbours.
}

\seealso{
\code{\link{squareCounts}}
}

\author{Aaron Lun}

\examples{
hic.file <- system.file("exdata", "hic_sort.bam", package="diffHic")
cuts <- readRDS(system.file("exdata", "cuts.rds", package="diffHic"))
param <- pairParam(cuts)

# Setting up the parameters
fout <- "output"
invisible(preparePairs(hic.file, param, fout))
data <- squareCounts(fout, param, width=50, filter=1L)

# Collating neighbor counts.
countNeighbors(data, flank=1)
countNeighbors(data, flank=2)
countNeighbors(data, flank=5)
}

\references{
Mizuguchi T et al. (2014). Cohesin-dependent globules and heterochromatin shape 3D genome architecture in S. pombe. \emph{Nature}. % For the diagonal stuff.
}
