\alias{enrichedGap}
\name{enrichedGap}

\title{Compute local enrichment}
\description{Calculate the log-fold increase in abundance for each bin pair against its local neighbourhood.}

\usage{
enrichedGap(data, width, flank=5, trend=c("global", "none", "chr"), 
    prior.count=2, span=0.3)
}

\arguments{
\item{data}{a \code{DIList} object containing bin pair counts, generated by \code{\link{squareCounts}}}
\item{width}{an integer scalar, specifying the size of each bin in base pairs}
\item{flank}{an integer scalar, specifying the number of bins to consider as the local neighbourhood}
\item{trend}{a character string indicating the type of distance correction to perform}
\item{prior.count}{a numeric scalar indicating the prior count to use in computing the log-fold increase}
\item{span}{a numeric scalar specifying the span for the loess fit during distance correction}
}

\details{
Consider the coordinates of the interaction space in terms of bins, and focus on any particular bin pair (named here as the target bin pair).
This target bin pair is characterized by six neighbourhood regions; four quadrants and two axial segments.
Each quadrant is a square with side lengths equal to \code{flank}, positioned such that the target bin pair lies at the corner of each quadrant.
Each axial segment is a rectangle of dimensions \code{(1, flank*2+1)}, running parallel to one of the axes and containing the target bin pair at the center.
Of course, the target bin pair itself is not included in either neighbourhood.

% Axial segments might be better than separate up/down/left/right neighbourhoods, as those run the risk of
% being too aggressive if the peak spills over into those neighbourhoods. Spillover is diluted a bit more
% when you consider up/down together or left/right together, which reduces the problem.

For a target bin pair in \code{data}, the mean abundance is computed for each of its surrounding neighbourhoods.
This is defined as the mean of the (distance-adjusted) counts for all constituent bin pairs in that neighbourhood (average counts are used for multiple libraries).
The local background for the target bin pair is defined as the maximum mean abundance across all neighbourhoods.
The enrichment value is then defined as the the difference between the target bin pair's abundance and its local background.
The idea is that bin pairs with high enrichments are likely to represent punctate interactions between clearly defined loci.
Selecting for high enrichments can then select for these peak-like features in the interaction space.

% If a peak is outside a TAD (but still within flank distance), then the TAD will be partially included in the 
% background. This could inflate the background abundance and prevent calling of the peak. To avoid this, you'd
% have to be able to specify the boundaries of the structures in the interaction space, which is complicated.

The maximizing strategy is designed to mitigate the effects of structural features.
Quadrants will capture the high interaction intensity within TADs or compartments, while the axial segments will capture any bands in the interaction space.
The abundance will be high for any neighbourhood that captures a high-intensity feature, as the average counts will be high for all bin pairs within these features.
This will then be chosen as the maximum during calculation of enrichment values.
Otherwise, the background will be diluted by low-intensity space for bin pairs on the edge of the features.
This results in spuriously high enrichment values.

Some distance adjustment is necessary to account for the decrease in the abundance with increases in the distance between interacting bins.
This is done by fitting a loess curve to the log-distance against the average abundance, for all intra-chromosomal bin pairs.
The log-average abundance for each bin pair is then decreased by the fitted value, to obtain a distance-adjusted log-average count.
The trend can be fitted to all intra-chromosomal bin pairs (if \code{trend="global"}) or for each chromosome separately (if \code{"chr"}).
Alternatively, no adjustment will be performed if \code{"none"}.
Obviously, this is the case for inter-chromosomal bin pairs.
}

\value{
A numeric vector containing the log-fold increase for each bin pair in \code{data}.
}

\author{
Aaron Lun
}

\seealso{
\code{\link{squareCounts}}
}

\examples{
hic.file <- system.file("exdata", "hic_sort.bam", package="diffHic")
cuts <- readRDS(system.file("exdata", "cuts.rds", package="diffHic"))
param <- pairParam(fragments=cuts)

# Setting up the parameters
fout <- "output.h5"
invisible(preparePairs(hic.file, param, file=fout))

# Collating to count combinations.
data <-	squareCounts(fout, param, filter=1, width=40)
head(enrichedGap(data))

head(enrichedGap(data, width=40))

head(enrichedGap(data, width=40, flank=5))
head(enrichedGap(data, width=40, flank=1))

head(enrichedGap(data, trend="none"))
}

\references{
Rao S et al. (2014). A 3D map of the human genome at kilobase resolution reveals principles of chromatin looping. \emph{Cell}. 159, 1665-1690.
}
