\alias{enrichedGap}
\name{enrichedGap}

\title{Compute local enrichment}
\description{Calculate the log-fold increase in abundance for each bin pair against its local neighbourhood.}

\usage{
enrichedGap(data, width, flank=3, trend=c("global", "none", "chr"), 
    prior.count=2, span=0.3)
}

\arguments{
\item{data}{a \code{DIList} object containing bin pair counts, generated by \code{\link{squareCounts}}}
\item{width}{an integer scalar, specifying the size of each bin in base pairs}
\item{flank}{an integer scalar, specifying the number of bins to consider as the local neighbourhood}
\item{trend}{a character string indicating the type of distance correction to perform}
\item{prior.count}{a numeric scalar indicating the prior count to use in computing the log-fold increase}
\item{span}{a numeric scalar specifying the span for the loess fit during distance correction}
}

\details{
For a target bin pair in \code{data}, this function will compute the mean abundance for each of 8 neighbourhoods, i.e., quadrants or extensions.
Each quadrant is a square with side lengths equal to \code{flank}, positioned such that the target bin pair lies at the corner of each quadrant.
In contrast, each extension is a rectangle with dimensions of \code{(1, flank)} or \code{(flank, 1)}, where the target bin pair makes up the first or last of that rectangle.
For each neighbourhood, the mean abundance is defined as the mean of the (distance-adjusted) average counts for all constituent bin pairs in that quadrant, excluding the target bin pair.
The local background is defined as the maximum mean abundance across all neighbourhoods.
The enrichment value is then defined as the the difference between the bin pair abundance and its local background.

The idea is that bin pairs with high enrichments are likely to represent punctate interactions between clearly defined loci.
Selecting for high enrichments can then select for these peak-like features in the interaction space.
The maximizing strategy is designed to mitigate the effects of structural features (e.g., TADs, compartments, banding effects).
The local background of at least one neighbourhood will be high for all bin pairs within these features.
Otherwise, the background will be diluted by low-intensity space for bin pairs on the edge of the features.
This results in spuriously high enrichment values, such that filtering on enrichments will be misleading.

Some distance adjustment is necessary to account for the decrease in the abundance with increases in the distance between interacting bins.
This is done by fitting a loess curve to the log-distance against the average abundance, for all intra-chromosomal bin pairs.
The log-average abundance for each bin pair is then decreased by the fitted value, to obtain a distance-adjusted log-average count.
The trend can be fitted to all intra-chromosomal bin pairs (if \code{trend="global"}) or for each chromosome separately (if \code{"chr"}).
Alternatively, no adjustment will be performed if \code{"none"}.
Obviously, this is the case for inter-chromosomal bin pairs.
}

\value{
A numeric vector containing the log-fold increase for each bin pair in \code{data}.
}

\author{
Aaron Lun
}

\seealso{
\code{\link{squareCounts}}
}

\examples{
hic.file <- system.file("exdata", "hic_sort.bam", package="diffHic")
cuts <- readRDS(system.file("exdata", "cuts.rds", package="diffHic"))
param <- pairParam(fragments=cuts)

# Setting up the parameters
fout <- "output.h5"
invisible(preparePairs(hic.file, param, file=fout))

# Collating to count combinations.
data <-	squareCounts(fout, param, filter=1, width=40)
head(enrichedGap(data))

head(enrichedGap(data, width=40))

head(enrichedGap(data, width=40, flank=5))
head(enrichedGap(data, width=40, flank=1))

head(enrichedGap(data, trend="none"))
}

\references{
Rao S et al. (2014). A 3D map of the human genome at kilobase resolution reveals principles of chromatin looping. \emph{Cell}. 159, 1665-1690.
}
